Add HasBatchSender/Receiver traits
