+++ crates/relayer-framework/src/full/batch/impls/mod.rs
@@ -1,1 +1,1 @@
-pub mod channel;
+pub mod relay;
