+++ crates/relayer-framework/src/full/batch/traits/mod.rs
@@ -1,1 +1,2 @@
 pub mod channel;
+pub mod types;
